templates/churn/pyproject.toml
[project]
name = "churn"
version = "0.1.0"
description = "Churn prediction template."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "devtools>=0.12.2",
    "keras>=3.6.0",
    "keras-nlp>=0.0.2",
    "matplotlib>=3.9.2",
    "numpy>=1.26.4",
    "polars>=1.9.0",
    "ruff>=0.6.9",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "tensorflow>=2.17.0",
    "tqdm>=4.66.5",
    "python-dotenv>=1.0.1",
    "fastapi[all]>=0.115.2",
    "google-cloud-storage>=2.18.2",
]

[tool.mypy]
plugins = 'pydantic.mypy'
strict = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
disallow_untyped_decorators = false
disallow_subclassing_any = false
ignore_missing_imports = true
follow_imports = 'skip'
exclude = [
    "venv",
    ".venv",
    "alembic",
    "app/dbs/sql_gen/gen/db.py"
]

[tool.ruff]
target-version = "py312"
exclude = [
    "alembic",
    "app/dbs/sql_gen/gen/db.py",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
unfixable = [
    "F401", # unused imports
]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
